fix: 修完阶段 2 独立评审的全部必须修项 + 大部分建议改项
评审（一个空 agent，只读、自己跑测试与接口）给出 5 条必须修、12 条建议改。
全部必须修已修，建议改里除「真机手机端验收」外的都处理了。

必须修：
1. 出勤率按**姓名**去重 → 同名学生会漏算（两个「张伟」都旷课只算 1 人，出勤率偏高）。
   改为按 student_id 去重（点名按 studentId 提交，同名能同时进来，是界面可达路径）。
   新增用例：同名两人 + 另一人全旷课 → absent=3。
2. 编辑作业会把「应交人数」快照重置成当前班级人数（PATCH 时钩子拿到的是空值，
   于是每次更新都重新快照），与 models/homework.py 和变更登记表 #3 的承诺相反。
   改为只在**创建**时快照；更新时留空保持原值 —— 顺便挡住把 total 写成 NULL 的路径。
3. `unsubmitted_names` 被声明为可排序但它是个派生属性，点一下表头 = AttributeError 500，
   而且前端会把那个排序键记进 state，这个页面再也刷不出列表；导出还会把 500 的 JSON
   当 xlsx 下载。修：`sortable=False`，并补一条**结构性**看门测试 ——
   声明可排序/可搜索/可筛选的键必须是真实列或 JSON 字段。
4. 「手动覆盖时提示差异」没做到。修：`compute_rate` 移到 models（models 不能 import
   services），新增派生值 `rate_auto`（按名单算，不落库），列表与导出里与生效值并排，
   作业页顶部再给一条不一致提醒（要求两个数并存时，别人知道该信哪个）。
5. 阶段 2 验收里三条在代码里不存在。修：实现 `GET /students/name-conflicts`
   （档案页顶部提示同名分组 + 可操作的处理建议）；「一生一档」「欠交次数」「首页待收」
   是**有意的推迟**（阶段 5 页面范围），已在 02 与 06 里写明推迟理由与验收方式。
   顺带发现：学号冲突结构上不可能（`uq_students_class_sno ... WHERE sno <> ''`），
   所以不需要「学号冲突报告」，靠唯一约束挡住即可 —— 用测试钉住这个结论。

建议改：
- 测试之间互相污染（session 级 client + 不回滚的 db_session）→ 补用例级业务表清理。
  加完立刻抓出一个**靠数据污染才通过**的用例（班里 0 人时它断言提交率 100%），已改成精确断言。
- `openModal({wide})` 是空参数（CSS 里没有 .modal.wide）→ 补样式，成绩录入表不再挤在 640px 里
- 同一次提交里同一格出现两次 → 撞唯一约束 500；改为按后面的为准
- `absent=bool("否")` 是 True → 改用 is_truthy
- 出勤重名报错让老师「用学号区分」，但那张表单只收姓名 → 改成做得到的出路（改用点名 / 去档案改名）
- 成绩页在没有考试时三个动作静默无反应 → 明确提示该怎么办
- 注释与行为不符（残留成绩行「录入表会照实显示」其实被过滤掉）→ 改成事实描述
- `previous_total` 标注 int 实为 float、测试名与断言不符 → 一并订正
- 文档同步：rate_mode 的中文取值写进 README §3.2 与 04

测试 145 → 164，100 个文件全部 ≤600 行。

+++ backend/app/services/score_stats.py
@@ -352,8 +352,9 @@ def build_report(session: Session, exam: Exam, *, include_previous: bool = True)
         result = by_student.get(score.student_id)
         stat = subject_index.get(score.subject)
         if result is None or stat is None:
-            # 科目被移出这场考试、或学生已不在本班时可能留下这样的行。
-            # 统计忽略它，但录入表会照实显示，让人看到并处理 —— 不在这里静默计入
+            # 这种行**通过接口造不出来**：`set_subjects` 拒绝移出已有成绩的科目，
+            # `save_cells` 拒绝不在本场科目里的分数。这里是防手工改库的兜底 ——
+            # 统计时忽略，不静默把它算进某个学生或某科
             continue
         if score.absent:
             result.absent.append(score.subject)
@@ -418,7 +419,7 @@ def _apply_ranks(rows: list[StudentResult]) -> None:
     """
     ranked = sorted((row for row in rows if row.took_part), key=lambda row: -row.total)
     rank = 0
-    previous_total: int | None = None
+    previous_total: float | None = None
     for index, row in enumerate(ranked, start=1):
         if previous_total is None or row.total < previous_total:
             rank = index  # 分数降了才占新名次，同分沿用上一个名次
